% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additive_summary.R
\name{additive_summary}
\alias{additive_summary}
\title{additive_summary}
\usage{
additive_summary(
  summaryCall,
  fhatSamples,
  fhat = rowMeans(fhatSamples),
  df = NULL,
  alpha = 0.05,
  ribbonFill = "grey50",
  meta = NA,
  verbose = FALSE
)
}
\value{

}
\description{
.. content for description{} (no empty lines) ..
}
\details{
.. content for details{} ..
}
\author{
Spencer Woody
}
