% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additive_summary.R
\name{additive_summary_local}
\alias{additive_summary_local}
\title{additive_summary}
\usage{
additive_summary_local(
  summaryCall,
  fhatSamples,
  fhat = rowMeans(fhatSamples),
  df = NULL,
  grid_size = 100,
  alpha = 0.05,
  ribbonFill = "grey50",
  return_samples = TRUE,
  meta = NA,
  verbose = FALSE
)
}
\arguments{
\item{summaryCall}{A gam fomula for the additive summary to be
computed.  Should be in the form of fhat ~ s(x1) + s(x2) + ....
See ?mgcv::formula.gam}
}
\value{

}
\description{
.. content for description{} (no empty lines) ..
}
\details{
.. content for details{} ..
}
\author{
Spencer Woody
}
