% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additive_summary_plot.R
\name{additive_summary_plot}
\alias{additive_summary_plot}
\title{Plot an additive summary}
\usage{
additive_summary_plot(additive_summary, ribbonFill = "grey80")
}
\arguments{
\item{additive_summary}{Output of additive_summary}

\item{ribbonFill}{An optional argument for the color of the credible interval}
}
\description{
Plot additive summary
}
\details{
Plots the partial/adjusted effect curves from additive_summary
}
\author{
Spencer Woody
}
