% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additiveSummary1.R
\name{additiveSummary1}
\alias{additiveSummary1}
\title{Additive summary}
\usage{
additiveSummary1(
  gamcall,
  fhatSamples,
  fhat = rowMeans(fhatSamples),
  df = NULL,
  terms = NULL,
  alpha = 0.05
)
}
\arguments{
\item{gamcall}{}

\item{df}{}

\item{yhatSamples}{}

\item{yhat}{}
}
\value{

}
\description{
.. content for description{} (no empty lines) ..
}
\details{
.. content for details{} ..
}
\author{
Spencer Woody
}
