% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subgroups.R
\name{average_posterior}
\alias{average_posterior}
\title{subgroup_average_posterior}
\usage{
average_posterior(posterior_samples, weights = NULL)
}
\arguments{
\item{posterior_samples}{A matrix of posterior samples, with MCMC draws down the rows}

\item{weights}{Optional weights, for computing a weighted subgroup average}
}
\description{
Aggregate posterior draws by averaging
}
\details{
Aggregate posterior draws by averaging, re (weighted) average
}
