% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_smlePossum.R
\name{cv_smlePossum}
\alias{cv_smlePossum}
\title{Cross-validated observed-data log-likelihood}
\usage{
cv_smlePossum(
  fold,
  Y,
  offset = NULL,
  X_unval,
  X_val,
  Z = NULL,
  Validated = NULL,
  Bspline = NULL,
  data,
  theta_pred = NULL,
  initial_lr_params = "Zero",
  TOL = 1e-04,
  MAX_ITER = 1000
)
}
\arguments{
\item{fold}{Column name with the assigned fold for cross-validation.}

\item{Y}{Column name with the outcome}

\item{offset}{(Optional) Column name with the offset for \code{Y}. Default is \code{offset = NULL} for no offset}

\item{X_unval}{Column name(s) with the unvalidated covariates.  If \code{X_unval} and \code{X_val} are \code{null}, all covariates are assumed to be error-free.}

\item{X_val}{Column name(s) with the validated covariates. If \code{X_unval} and \code{X_val} are \code{null}, all covariates are assumed to be error-free.}

\item{Z}{(Optional) Column name(s) with additional error-free covariates.}

\item{Validated}{Column name with the validation indicator. The validation indicator can be defined as \code{Validated = 1} or \code{TRUE} if the subject was validated and \code{Validated = 0} or \code{FALSE} otherwise.}

\item{Bspline}{Vector of column names containing the B-spline basis functions.}

\item{data}{A dataframe with one row per subject containing columns: \code{Y_unval}, \code{Y}, \code{X_unval}, \code{X_val}, \code{Z}, \code{Validated}, and \code{Bspline}.}

\item{theta_pred}{Vector of columns in \code{data} that pertain to the covariates in the analysis model. The default assumes main effects of \code{X_val} and \code{Z} only.}

\item{initial_lr_params}{Initial values for parametric model parameters. Choices include (1) \code{"Zero"} (non-informative starting values) or (2) \code{"Complete-data"} (estimated based on validated subjects only)}

\item{TOL}{Tolerance between iterations in the EM algorithm used to define convergence.}

\item{MAX_ITER}{Maximum number of iterations allowed in the EM algorithm.}
}
\value{
scalar value of the function
}
\description{
This function returns the value of the observed-data log-likelihood for the SMLE based on cross-validation.
}
