% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subgroups.R
\name{difference_posterior}
\alias{difference_posterior}
\title{Title}
\usage{
difference_posterior(posterior_samples, ref_col = 1, remove_ref = TRUE)
}
\arguments{
\item{posterior_samples}{}

\item{ref_col}{}

\item{remove_ref}{}
}
\value{

}
\description{
Title
}
