% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamProjFun2.R
\name{gamProjectionFun}
\alias{gamProjectionFun}
\title{Projection}
\usage{
gamProjectionFun(
  gamcall,
  fhatSamples,
  fhat = rowMeans(fhatSamples),
  df = NULL,
  alpha = 0.05,
  ribbonFill = "grey50",
  meta = NA,
  verbose = FALSE
)
}
\value{

}
\description{
.. content for description{} (no empty lines) ..
}
\details{
.. content for details{} ..
}
\author{
Spencer Woody
}
