% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_additive_factor_posterior}
\alias{get_additive_factor_posterior}
\title{Return posterior samples of the coefficients on factor variables in a generalized additive posterior summary}
\usage{
get_additive_factor_posterior(possum_fit)
}
\arguments{
\item{possum_fit}{A posterior summary object (e.g. as returned by additive_summary)}
}
\value{
A matrix of posterior samples, with one column per coefficient
}
\description{
Return posterior samples of the coefficients on factor variables in a generalized additive posterior summary
}
