% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_additive_summary_Rsq}
\alias{get_additive_summary_Rsq}
\title{Return the posterior summary R^2 diagnostic}
\usage{
get_additive_summary_Rsq(possum_fit)
}
\arguments{
\item{possum_fit}{A posterior summary object (e.g. returned by additive_summary)}
}
\value{
A vector of posterior samples
}
\description{
Returns posterior samples of the R^2 diagnostic for psoterior summaries
}
