% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impBeaver.R
\name{impBeaver}
\alias{impBeaver}
\title{Multiple imputation for log-binomial regression problems with covariate measurement error
This function returns the multiple imputation (MI) estimates for the Poisson regression model with covariate measurement error from Lotspeich et al. (2023+)}
\usage{
impBeaver(
  imputation_formula,
  analysis_formula,
  data,
  adjMatrix = NULL,
  B = 1,
  seed = NULL
)
}
\arguments{
\item{imputation_formula}{imputation model formula (or coercible to formula) passed through to \code{lm()}, a formula expression as for other regression models. The response should be the error-prone version of the covariate.}

\item{analysis_formula}{analysis model formula (or coercible to formula) passed through to \code{glm()}, a formula expression as for other regression models. The response should be the log-binomial model outcome as \code{cbind(cases, controls)}.}

\item{data}{dataset containing at least the variables included in \code{imputation_formula} and \code{analysis_formula}.}

\item{adjMatrix}{(optional) adjacency matrix for observations in \code{data} to be passed through to \code{spaMM::fitme()}. If \code{adjMatrix} is not \code{NULL} (the default), then a spatial random effect should be included in \code{analysis_formula}.}

\item{B}{desired number of imputations. Default is \code{B = 1}, which is single imputation.}

\item{seed}{(optional) random seed to use for reproducibility of random draws. Default is \code{seed = NULL}, which does not reset the random seed inside the function.}
}
\value{
dataframe with final coefficient and standard error estimates for the analysis model, pooled according to Rubin's rules.
}
\description{
Multiple imputation for log-binomial regression problems with covariate measurement error
This function returns the multiple imputation (MI) estimates for the Poisson regression model with covariate measurement error from Lotspeich et al. (2023+)
}
