% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impPossum_data.R
\name{impPossum_data}
\alias{impPossum_data}
\title{Multiple imputation for Poisson regression problems with covariate measurement error
This function returns the multiply imputed datasets}
\usage{
impPossum_data(imputation_formula, analysis_formula, data, B = 1, seed = NULL)
}
\arguments{
\item{imputation_formula}{imputation model formula (or coercible to formula) passed through to \code{lm()}, a formula expression as for other regression models. The response should be the error-prone version of the covariate.}

\item{analysis_formula}{analysis model formula (or coercible to formula) passed through to \code{glm()}, a formula expression as for other regression models. The response should be the Poisson model outcome, and, if needed, the offset can be included in this formula using the \code{offset()} function.}

\item{data}{dataset containing at least the variables included in \code{imputation_formula} and \code{analysis_formula}.}

\item{B}{desired number of imputations. Default is \code{B = 1}, which is single imputation.}

\item{seed}{(optional) random seed to use for reproducibility of random draws. Default is \code{seed = NULL}, which does not reset the random seed inside the function.}
}
\value{
dataframe with final coefficient and standard error estimates for the analysis model, pooled according to Rubin's rules.
}
\description{
Multiple imputation for Poisson regression problems with covariate measurement error
This function returns the multiply imputed datasets
}
