% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlePossum_helpers.R
\name{mle_loglik}
\alias{mle_loglik}
\title{Observed-data log-likelihood for the sieve maximum likelihood estimator (SMLE)}
\usage{
mle_loglik(
  Y = NULL,
  offset = NULL,
  X_unval = NULL,
  X = NULL,
  Z = NULL,
  comp_dat_val,
  comp_dat_unval,
  beta,
  eta,
  noFN
)
}
\arguments{
\item{Y}{Column name with the outcome}

\item{offset}{(Optional) Column name with the offset for \code{Y}. Default is \code{offset = 1}, no offset}

\item{X_unval}{Column(s) with the unvalidated covariates}

\item{X}{Column(s) with the validated covariates}

\item{Z}{(Optional) Column(s) with additional error-free covariates}

\item{comp_dat_val}{Dataset containing rows for validated subjects' data (a matrix)}

\item{comp_dat_unval}{Augmented dataset containing rows for each combination of unvalidated subjects' data with values from Phase II (a matrix)}

\item{beta}{Parameters for the analysis model (a column vector)}

\item{eta}{Parameters for the misclassification model (a column vector)}

\item{noFN}{logical, if \code{noFN = FALSE} (the default), then it is assumed that there can be both false positives and false negatives in the error-prone exposure. If \code{noFN = TRUE}, the error mechanism is restricted to only false positives.}
}
\value{
Scalar value of the function
}
\description{
This function returns the value of the observed-data log-likelihood (equation (#) in Lotspeich et al. (2023+))
for a given dataset and parameter values \code{beta} and \code{p}.
}
