% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phiGamma.R
\name{phiGamma}
\alias{phiGamma}
\title{Inflation of stdev}
\usage{
phiGamma(y, gamma, sigma)
}
\arguments{
\item{y}{}

\item{gamma}{}

\item{sigma}{}
}
\value{

}
\description{
Inflation of sigma
}
\details{
.. content for \details{} ..
}
\author{
Spencer Woody
}
