% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possum1.R
\name{possum1}
\alias{possum1}
\title{possum1}
\usage{
possum1(
  summaryCall,
  fhatSamples,
  fhat = rowMeans(fhatSamples),
  df = NULL,
  alpha = 0.05,
  ribbonFill = "grey50",
  meta = NA,
  verbose = FALSE
)
}
\value{

}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Spencer Woody
}
