% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possum_plot1.R
\name{possum_plot1}
\alias{possum_plot1}
\title{Plot}
\usage{
possum_plot1(possum1)
}
\value{

}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Spencer Woody
}
