% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possum_triangle1.R
\name{possum_triangle1}
\alias{possum_triangle1}
\title{Triangle}
\usage{
possum_triangle1(possum1)
}
\value{

}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Spencer Woody
}
