% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsqGamma.R
\name{rsqGamma}
\alias{rsqGamma}
\title{A function for computing the summary R-squared}
\usage{
rsqGamma(gamma, fhat)
}
\arguments{
\item{gamma}{fitted values of summary (vector or matrix)}

\item{fhat}{fitted values from regression function (vector or
matrix)}
}
\description{
Description
}
\details{
Details
}
\author{
Spencer Woody
}
