% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smlePossum.R
\name{smlePossum}
\alias{smlePossum}
\title{Sieve maximum likelihood estimator (SMLE) for two-phase Poisson regression problems with covariate measurement error
This function returns the sieve maximum likelihood estimators (SMLE) for the Poisson regression model with covariate measurement error from Lotspeich et al. (2023+)}
\usage{
smlePossum(
  Y,
  offset = NULL,
  X_unval,
  X,
  Z = NULL,
  data,
  beta_pred = NULL,
  beta0 = "Zero",
  h_N_scale = 1,
  noSE = FALSE,
  TOL = 1e-04,
  MAX_ITER = 1000
)
}
\arguments{
\item{Y}{Column name with the outcome}

\item{offset}{(Optional) Column name with the offset for \code{Y}. Default is \code{offset = NULL} for no offset.}

\item{X_unval}{Column name(s) with the unvalidated covariates.}

\item{X}{Column name(s) with the validated covariates.}

\item{Z}{(Optional) Column name(s) with additional error-free covariates.}

\item{data}{A dataframe with one row per subject containing columns: \code{Y}, \code{X_unval}, \code{X}, \code{Z}, \code{Validated}, and \code{Bspline}.}

\item{beta_pred}{Vector of columns in \code{data} that pertain to the covariates in the analysis model. The default assumes main effects of \code{X} and \code{Z} only.}

\item{beta0}{Initial values for parametric model parameters. Choices include (1) \code{"Zero"} (non-informative starting values, the default) or (2) \code{"Complete-data"} (estimated based on validated subjects only)}

\item{h_N_scale}{Size of the perturbation used in estimating the standard errors via profile likelihood. If none is supplied, default is \code{h_N_scale = 1}.}

\item{noSE}{Indicator for whether standard errors are desired. Defaults to \code{noSE = FALSE}.}

\item{TOL}{Tolerance between iterations in the EM algorithm used to define convergence.}

\item{MAX_ITER}{Maximum number of iterations allowed in the EM algorithm.}
}
\value{
\item{coeff}{dataframe with final coefficient and standard error estimates (where applicable) for the analysis model.}
\item{ppv}{final positive predictive value estimate.}
\item{vcov}{variance-covariance matrix for \code{coeff} (where applicable).}
\item{converged}{indicator of EM algorithm convergence for parameter estimates.}
\item{se_converged}{indicator of standard error estimate convergence.}
\item{converged_msg}{(where applicable) description of non-convergence.}
\item{iterations}{number of iterations completed by EM algorithm to find parameter estimates.}
\item{od_loglik_at_conv}{value of the observed-data log-likelihood at convergence.}
}
\description{
Sieve maximum likelihood estimator (SMLE) for two-phase Poisson regression problems with covariate measurement error
This function returns the sieve maximum likelihood estimators (SMLE) for the Poisson regression model with covariate measurement error from Lotspeich et al. (2023+)
}
