% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_linear_summary.R
\name{sparse_linear_summary}
\alias{sparse_linear_summary}
\title{Sparse linear summaries}
\usage{
sparse_linear_summary(
  X,
  fhatmat = X \%*\% betaSamples,
  betaSamples,
  sigma2Samples = NA,
  adaptive = TRUE,
  varnames = NA,
  alpha = 0.05,
  ...
)
}
\arguments{
\item{X}{N \times p design matrix}

\item{fhatmat}{N \times NMC matrix of posterior draws of the function f, where N is the number of observations and NMC is the number of Monte Carlo posterior samples. The user must specify fhatmat OR betaSamples}

\item{betaSamples}{p \times NMC matrix of posterior draws of the coefficients for the (generalized) linear model}

\item{sigma2Samples}{Optional vector of posterior samples for the residual variance (for a linear model)}

\item{adaptive}{if TRUE (default), use adaptive lasso, weighting by the posterior mean. See Hahn and Carvalho (2015)}

\item{varnames}{Optional vector of variable names}

\item{alpha}{Return the alpha/2 and 1-alpha/2 posterior credible intervals for the summary}

\item{...}{other arguments, e.g., to glmnet}
}
\description{
Sparse linear summary
}
\details{
Compute a sparse linear summary of a nonparametric regression model or high-dimensional (generalized) linear model
}
\author{
Spencer Woody
}
