% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseLinearProj.R
\name{sparseLinearProj}
\alias{sparseLinearProj}
\title{Sparse Linear Projection}
\usage{
sparseLinearProj(
  X,
  y,
  betaSamples,
  sigma2Samples,
  adaptive = TRUE,
  varnames = NA,
  alpha = 0.05,
  ...
)
}
\arguments{
\item{X}{}

\item{y}{}

\item{betaSamples}{}

\item{sigma2Samples}{}

\item{adaptive}{if TRUE (default), use adaptive lasso, weighting
by the postserior mean}

\item{varnames}{}

\item{alpha}{}

\item{...}{other arguments, e.g., to glmnet}
}
\value{

}
\description{
Description
}
\details{
Details
}
\author{
Spencer Woody
}
