% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseLinearSummary.R
\name{sparseLinearSummary}
\alias{sparseLinearSummary}
\title{Sparse linear summarie}
\usage{
sparseLinearSummary(
  X,
  fhatmat,
  y,
  sigma2Samples,
  adaptive = TRUE,
  varnames = NA,
  alpha = 0.05,
  ...
)
}
\arguments{
\item{X}{}

\item{fhatmat}{}

\item{y}{}

\item{sigma2Samples}{}

\item{adaptive}{if TRUE (default), use adaptive lasso, weighting
by the postserior mean}

\item{varnames}{}

\item{alpha}{}

\item{...}{other arguments, e.g., to glmnet}
}
\value{

}
\description{
Description
}
\details{
Details
}
\author{
Spencer Woody
}
