% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subgroups.R
\name{subgroup_average_posterior}
\alias{subgroup_average_posterior}
\title{subgroup_average_posterior}
\usage{
subgroup_average_posterior(posterior_samples, groups, weights = NULL)
}
\arguments{
\item{posterior_samples}{A matrix of posterior samples, with MCMC draws down the rows}

\item{groups}{A factor defining the desired subgroups. Must have length ncol(posterior_samples)}

\item{weights}{Optional weights, for computing a weighted subgroup average}
}
\value{
A matrix of size (number of MCMC samples) x (number of subgroups) 
containing posterior draws of the subgroup averages
}
\description{
Aggregate posterior draws into subgroups by a factor
}
\details{
Aggregate posterior draws into subgroups by a factor and return a new
matrix of posterior draws for the subgroup (weighted) average
}
