% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangle.R
\name{triangle}
\alias{triangle}
\title{Triangles}
\usage{
triangle(x, fsamples, num_quantiles = 100)
}
\description{
.. content for description{} (no empty lines) ..
}
\details{
.. content for details{} ..
}
\author{
Spencer Woody
}
