% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{U}
\alias{U}
\title{score}
\usage{
U(beta, x, y)
}
\arguments{
\item{beta}{a numeric vector of parameters}

\item{x}{a numeric vector of covariates}

\item{y}{a numeric vector of outcomes}
}
\value{
a scalar
}
\description{
score
}
