% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additive_summary_triangle_plot.R
\name{additive_summary_triangle_plot}
\alias{additive_summary_triangle_plot}
\title{Triangle}
\usage{
additive_summary_triangle_plot(possum1)
}
\value{

}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Spencer Woody
}
