% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subgroups.R
\name{difference_posterior}
\alias{difference_posterior}
\title{Difference_posterior}
\usage{
difference_posterior(posterior_samples, ref_col = 1, remove_ref = TRUE)
}
\arguments{
\item{posterior_samples}{Set of posterior samples}

\item{ref_col}{Differences are taken as (each column) - (ref_col)}

\item{remove_ref}{Remove the column of zeros corresponding to the reference column}
}
\value{

}
\description{
Difference_posterior
}
