% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smlePossum_helpers.R
\name{smle_loglik}
\alias{smle_loglik}
\title{Observed-data log-likelihood for the sieve maximum likelihood estimator (SMLE)}
\usage{
smle_loglik(
  Y = NULL,
  offset = NULL,
  X_unval = NULL,
  X_val = NULL,
  Z = NULL,
  Bspline = NULL,
  comp_dat_val,
  comp_dat_unval,
  theta_pred,
  theta,
  p
)
}
\arguments{
\item{Y}{Column name with the outcome}

\item{offset}{(Optional) Column name with the offset for \code{Y}. Default is \code{offset = 1}, no offset}

\item{X_unval}{Column(s) with the unvalidated covariates}

\item{X_val}{Column(s) with the validated covariates}

\item{Z}{(Optional) Column(s) with additional error-free covariates}

\item{Bspline}{Vector of columns containing the B-spline basis functions}

\item{comp_dat_val}{Dataset containing rows for validated subjects' data (a matrix)}

\item{comp_dat_unval}{Augmented dataset containing rows for each combination of unvalidated subjects' data with values from Phase II (a matrix)}

\item{theta_pred}{Vector of columns in \code{data} that pertain to the covariates in the analysis model.}

\item{theta}{Parameters for the analysis model (a column vector)}

\item{p}{B-spline coefficients for the approximated covariate error model (a matrix)}
}
\value{
Scalar value of the function
}
\description{
This function returns the value of the observed-data log-likelihood (equation (#) in Lotspeich et al. (2023+))
for a given dataset and parameter values \code{theta} and \code{p}.
}
