% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseLinearProj.R
\name{sparseLinearProj}
\alias{sparseLinearProj}
\title{Sparse Linear Projection (a la Hahn & Carvalho, 2015)}
\usage{
sparseLinearProj(
  X,
  y,
  betaSamples,
  sigma2Samples,
  adaptive = TRUE,
  varnames = NA,
  alpha = 0.05,
  ...
)
}
\arguments{
\item{adaptive}{if TRUE (default), use adaptive lasso, weighting
by the postserior mean}

\item{...}{other arguments, e.g., to glmnet}
}
\value{

}
\description{
This is like DSS
}
\details{
Details
}
\author{
Spencer Woody
}
